My solution versus Angela Yu's solution
My solution is the index.html file, hers is solution.html.

The difference I took was to make the gaps themselves their own grids, and what she did, to simplify, was to make the background color of the entire container black and put gaps. I wasn't sure exactly how she came to that at the time, so instead, I just made everything grids. Either way, we both came to the same solution.